requirements.txt
hydra-core==1.3.2
pytorch-lightning==2.0.2
torch==2.0.1
tensorboard==2.13.0
torchmetrics==0.11.4
torchvision==0.15.2
-e .
